requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1
PyMuPDF==1.26.0
python-docx==1.1.2
requests==2.32.3
qdrant-client==1.14.2
python-multipart==0.0.20
